test: fix session_id errors in test runs
Upgrade to Drupal 9 and PhpUnit 9 broke our existing site tests. The tests did work but it was a lucky coincidence it worked.

The problem started because of this change: https://www.drupal.org/node/3006306. It seems innocuous and it is. The problem was that we were directly calling views_get_view_result in a DTT based test. When we switch to checking the HTTP response of the page generated by the view, the tests works.

A more proper fix might appear in https://www.drupal.org/project/flag/issues/3221615 but for now, the tests pass.