fix: broken GitHub scanning
This broke sometime back in November and the fix was staring right in my face, but I didn't catch it. The change resulted in `client` always being set to NULL (because the authenticate method returns NULL). This PR fixes it by breaking the assignment and function call on different lines.